Criando opreacao de instanciacao

+++ src/main/java/com/juliana/cursomcj/repositories/CategoriaRepository.java
@@ -6,6 +6,9 @@ import org.springframework.stereotype.Repository;
 import com.juliana.cursomcj.domain.Categoria;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer>{
+public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
 
+	
+    // Aqui você pode adicionar métodos customizados se precisar, como:
+    // List<Categoria> findByNomeContaining(String nome);
 }
