+++ src/main/java/com/juliana/cursomcj/resources/CategoriaResource.java
@@ -1,8 +1,5 @@
 package com.juliana.cursomcj.resources;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -23,7 +20,7 @@ public class CategoriaResource {
 	@Autowired
 	private CategoriaService service;
 	@RequestMapping(value = "/{id}",method=RequestMethod.GET)
-	public ResponseEntity<?> find(@PathVariable Integer id) {
+	public ResponseEntity<?> find(@PathVariable Long id) {
 		Categoria obj = service.buscar(id);
 		return ResponseEntity.ok().body(obj);
